Delete prompt confirmation

+++ src/app/home/home.component.ts
@@ -1,10 +1,11 @@
-import { Component, OnInit } from '@angular/core';
-import { Router } from "@angular/router";
-import { PromptsService } from "../providers/prompts.service";
-import { Observable } from "rxjs";
-import { PromptDto } from "../providers/dto/prompt.dto";
+import {Component, OnInit} from '@angular/core';
+import {Router} from "@angular/router";
+import {PromptsService} from "../providers/prompts.service";
+import {PromptDto} from "../providers/dto/prompt.dto";
 import {MatSnackBar} from "@angular/material/snack-bar";
 import {UserService} from "../providers/user.service";
+import {ConfirmDialogComponent} from "../core/components/confirm-dialog/confirm-dialog.component";
+import {MatDialog} from "@angular/material/dialog";
 
 @Component({
   selector: 'app-home',
@@ -14,16 +15,27 @@ import {UserService} from "../providers/user.service";
 export class HomeComponent implements OnInit {
 
   constructor(private router: Router,
-    private snackBarService: MatSnackBar,
-    public userService: UserService,
-    private readonly promptsService: PromptsService) {
+              private dialog: MatDialog,
+              private snackBarService: MatSnackBar,
+              public userService: UserService,
+              private readonly promptsService: PromptsService) {
   }
 
   ngOnInit() {
   }
 
   delete(prompt: PromptDto) {
-    this.promptsService.deletePrompt(prompt.id).subscribe(() => this.deleteSuccess(prompt));
+    this.dialog.open(ConfirmDialogComponent,
+      {
+        width: '400px',
+        height: '200px',
+      }).afterClosed().subscribe((result) => this.onConfirmDialogClosed(result, prompt));
+  }
+
+  onConfirmDialogClosed(result: boolean, prompt: PromptDto) {
+    if (result) {
+      this.promptsService.deletePrompt(prompt.id).subscribe(() => this.deleteSuccess(prompt));
+    }
   }
 
   deleteSuccess(prompt: PromptDto) {
